move daemons unit tests to pytest
